backend/gateway-service/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
python-jose
websockets==15.0.1
python-multipart==0.0.21
prometheus-fastapi-intrumentator==7.1.0
pytest==9.0.2
pytest-asyncio==1.3.0
respx==0.22.0
pytest-cov==7.0.0
structlog==25.5.0